requirements.txt
Django==2.2.24
djangorestframework==3.11.2
pytz==2021.3
sqlparse==0.4.2
uWSGI==2.0.20
